[feature/#211] Comment entity, dto, service 구현

+++ resource-server/src/main/java/com/inhabas/api/domain/comment/repository/CommentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer>, CustomCommentRepository {
+public interface CommentRepository extends JpaRepository<Comment, Long>, CustomCommentRepository {
 
 }
